friday 16:00 12.05

+++ distanceFunctions/Hamming.py
@@ -2,7 +2,7 @@ def Hamming(u,v,type_values, hyperparams):
     distance = 0
     results = []
     for i in range(len(u)):
-        if int(v[i]) != int(u[i]):
+        if float(v[i]) != float(u[i]):
             results.append(1)
             distance += 1
         else:
